Refactor Quote Reset Time to improve the performance of swap
