+++ packages/extension-web-ui/src/components/Swap/index.tsx
@@ -3,3 +3,4 @@
 
 export { default as SwapRoute } from './SwapRoute';
 export { default as SwapTransactionBlock } from './SwapTransactionBlock';
+export { default as QuoteResetTime } from './QuoteResetTime';
